requirements-docker.txt
psycopg2
awesome-slugify==1.6.5
arrow==0.10.0
djangorestframework==3.5.3
Django==1.8.17
django-formtools==1.0
django-js-reverse==0.7.2
django-redis==4.6.0
evernote3
feedparser==5.2.1
github3.py==1.0.0a4
pocket==0.3.6
py-trello==0.7.0
pytz==2016.7
pytidylib==0.3.2
pypandoc==1.3.3
requests-oauthlib==0.7.0
twython==3.4.0
wallabag_api==1.1.0
flake8==3.2.1
todoist-python==7.0.13
pushbullet.py==0.10.0
instapush==0.1.2
python-taiga==0.8.6
